Main Menu COMPLETE
git ignore please work this time
