+++ Strange Space Spaceship Builder/Assets/Scenes/BuildMenu.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: f79968d9933b4044db8c9593483a50f4
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
