--- Strange Space Spaceship Builder/Assets/Scenes/MainMenuScript.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 649f875ca23a60c459fe8852444f1119
