+++ Strange Space Spaceship Builder/Library/LastSceneManagerSetup.txt
@@ -1,5 +1,5 @@
 sceneSetups:
-- path: Assets/Scenes/MainMenuScene.unity
+- path: Assets/Scenes/BuildMenu.unity
   isLoaded: 1
   isActive: 1
   isSubScene: 0
