+++ Strange Space Spaceship Builder/Library/StateCache/SceneView/8c/8cd7c613bf844de3b80696e27a479d5e.json
@@ -1,1 +1,1 @@
-﻿{"cameraMode":{"drawMode":0,"name":"Shaded","section":"Shading Mode"},"sceneLighting":true,"audioPlay":false,"sceneViewState":{"m_AlwaysRefresh":false,"showFog":true,"showSkybox":true,"showFlares":true,"showImageEffects":true,"showParticleSystems":true,"showVisualEffectGraphs":true,"m_FxEnabled":true},"in2DMode":false,"pivot":{"x":-1.5488150119781495,"y":9.448454856872559,"z":39.39430236816406},"rotation":{"x":0.12865397334098817,"y":0.001482869847677648,"z":-0.00019268043979536742,"w":0.9917022585868836},"size":16.81304168701172,"orthographic":false}
+﻿{"cameraMode":{"drawMode":0,"name":"Shaded","section":"Shading Mode"},"sceneLighting":true,"audioPlay":false,"sceneViewState":{"m_AlwaysRefresh":false,"showFog":true,"showSkybox":true,"showFlares":true,"showImageEffects":true,"showParticleSystems":true,"showVisualEffectGraphs":true,"m_FxEnabled":true},"in2DMode":false,"pivot":{"x":7.596966743469238,"y":3.6268680095672609,"z":-9.865691184997559},"rotation":{"x":0.023408187553286554,"y":0.0026946403086185457,"z":-0.00006334593490464613,"w":0.9997401833534241},"size":7.293606758117676,"orthographic":false}
