+++ Strange Space Spaceship Builder/Library/StateCache/SceneView/d0/d090dd8d83a43ec31f18a6a3a95d620e.json
@@ -1,0 +1,1 @@
+﻿{"cameraMode":{"drawMode":0,"name":"Shaded","section":"Shading Mode"},"sceneLighting":true,"audioPlay":false,"sceneViewState":{"m_AlwaysRefresh":false,"showFog":true,"showSkybox":true,"showFlares":true,"showImageEffects":true,"showParticleSystems":true,"showVisualEffectGraphs":true,"m_FxEnabled":true},"in2DMode":false,"pivot":{"x":-0.8548003435134888,"y":10.757709503173829,"z":39.848785400390628},"rotation":{"x":0.08211804181337357,"y":0.0038822386413812639,"z":-0.00032015025499276817,"w":0.9966302514076233},"size":28.612207412719728,"orthographic":false}
