+++ Strange Space Spaceship Builder/UserSettings/EditorUserSettings.asset
@@ -6,7 +6,13 @@ EditorUserSettings:
   serializedVersion: 4
   m_ConfigSettings:
     RecentlyUsedSceneGuid-0:
+      value: 5654505106055c030f0c58204375594441154829792c7f3175784f6bb4e16168
+      flags: 0
+    RecentlyUsedSceneGuid-1:
       value: 5a08575f5207595a0f5d59741173094444164f7d7d2a23317c7a4465bbe1646d
+      flags: 0
+    RecentlyUsedSceneGuid-2:
+      value: 05060d5f530d0d02545d5f2044730b4412154029757d7f6379714e32b7b0326c
       flags: 0
     vcSharedLogLevel:
       value: 0d5e400f0650
